fix: model hash is wrongly calculated causing rendering issues (only shown in debug build)
The state field is not initialized when calculating the model hash in FindModelFrame method. The hash is calculated based on data starting from type field up to hashnext, but the presence of the `state` field was causing the hash mismatch in debug build, causing sprite to be rendered instead of 3d model.

Since the state field is not used, it has been removed completely. Adding fields between type and hashnext might create future issues if these fields are not correctly initizialized before hashing